requirements.txt
beautifulsoup4
decord
numpy
pillow
playwright
PyYAML
requests
tqdm
